requirements.txt
flask
pillow
pyyaml
numpy
cnocr
onnxruntime